% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05b_deterministic_analysis_functions.R
\name{calculate_ce_out}
\alias{calculate_ce_out}
\title{Calculate cost-effectiveness outcomes}
\usage{
calculate_ce_out(l_params_all = load_all_params(), n_wtp = 1e+05)
}
\arguments{
\item{l_params_all}{List with all parameters of decision model}

\item{n_wtp}{Willingness-to-pay threshold to compute net benefits.}
}
\value{
A data frame with discounted costs, effectiveness and NMB for each strategy.
}
\description{
\code{calculate_ce_out} calculates costs and effects for a given vector of
parameters using a decision model. This function needs to be modified by the
users to fit their needs
}
