% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_simulation_model_functions.R
\name{check_transition_array}
\alias{check_transition_array}
\title{Check if transition array is valid}
\usage{
check_transition_array(a_P, n_states, n_t, err_stop = FALSE,
  verbose = FALSE)
}
\arguments{
\item{a_P}{A transition probability array.}

\item{n_states}{Number of health states.}

\item{n_t}{Number of cycles.}

\item{err_stop}{Logical variable to indicate stop process if error is found. 
Default = TRUE}

\item{verbose}{Logical variable to indicate print out of messages. 
Default = TRUE}
}
\value{
This function stops if transition probability array is not valid and shows 
what are the entries that are not valid
}
\description{
\code{check_transition_array} checks if transition probabilities are valid. 
This checks that each of the rows of the transition matrices sum to one and 
that each of its entries are in \[0, 1\].
}
