% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_simulation_model_functions.R
\name{decision_model}
\alias{decision_model}
\title{Decision Model}
\usage{
decision_model(l_params_all = load_all_params(), err_stop = FALSE,
  verbose = FALSE)
}
\arguments{
\item{l_params_all}{List with all parameters of decision model. By default, 
uses the base-case parameters generated by \code{\link{load_all_params}}}

\item{err_stop}{Logical variable to indicate stop process if error is found. 
Default = TRUE}

\item{verbose}{Logical variable to indicate print out of messages}
}
\value{
The transition probability array \code{a_P} and the cohort trace matrix
\code{m_M}.
}
\description{
\code{decision_model} implements the decision model.
}
