% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_calib_post_dist.R
\docType{data}
\name{m_calib_post}
\alias{m_calib_post}
\title{Posterior distribution of calibrated parameters}
\format{A matrix with 1000 rows and 3 variables:
\describe{
\item{p_S1S2}{Annual transition probability of disease progression (S1 to S2)}
\item{hr_S1}{Hazard rate ratio of death in S1 vs H}
\item{hr_S2}{Hazard rate ratio of death in S2 vs H}
}}
\usage{
m_calib_post
}
\description{
A dataset with the posterior distribution of the calibrated parameters of
the decision model.
}
\keyword{datasets}
