% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_all_cause_mortality.R
\docType{data}
\name{all_cause_mortality}
\alias{all_cause_mortality}
\title{Life table for the US in 2015}
\format{A data frame with 11 rows and 5 variables:
\describe{
\item{Year}{calendar year}
\item{Age}{age, in years}
\item{Female}{mortality rate for females}
\item{Male}{mortality rate for males}
\item{Total}{mortality rate for the total population}
}}
\usage{
all_cause_mortality
}
\description{
A dataset with all-cause mortality rates from US 2015 life table estimates
}
\keyword{datasets}
