% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_decision_model_functions.R
\name{check_transition_probability}
\alias{check_transition_probability}
\title{Check if transition array is valid}
\usage{
check_transition_probability(a_P, err_stop = FALSE, verbose = FALSE)
}
\arguments{
\item{a_P}{A transition probability array.}

\item{err_stop}{Logical variable to stop model run if set up as TRUE. Default = FALSE.}

\item{verbose}{Logical variable to indicate print out of messages.
Default = FALSE}
}
\value{
This function stops if transition probability array is not valid and shows
what are the entries that are not valid
}
\description{
\code{check_transition_probability} checks if transition probabilities are in [0, 1].
}
