% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_decision_model_functions.R
\name{decision_model}
\alias{decision_model}
\title{Decision Model}
\usage{
decision_model(l_params_all, err_stop = FALSE, verbose = FALSE)
}
\arguments{
\item{l_params_all}{List with all parameters of decision model}

\item{err_stop}{Logical variable to stop model run if set up as TRUE. Default = FALSE.}

\item{verbose}{Logical variable to indicate print out of messages. Default = FALSE}
}
\value{
The transition probability array and the cohort trace matrix.
}
\description{
\code{decision_model} implements the decision model used.
}
