% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cea_psa.R
\docType{data}
\name{df_cea_psa}
\alias{df_cea_psa}
\title{Cost-effectiveness results from probabilistic analysis}
\format{A \code{data.frame} with with 2 rows, ane per strategy and
5 variables:
\describe{
\item{Strategy}{Strategy name}
\item{Cost}{Cost per strategy}
\item{Effect}{QALYs per strategy}
\item{Inc_Cost}{Incremental cost}
\item{Inc_Effect}{Incremental QALYs}
\item{ICER}{Incremental cost-effectivenes ratio (ICER)}
\item{Status}{Domination status. ND, not dominated (i.e., on the
cost-effectivenes efficiency frontier); D, strongly dominated; d,
dominated by extension}
}}
\usage{
df_cea_psa
}
\description{
A dataset with cost and effectiveness outputs for each strategy.
}
\keyword{datasets}
