% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05a_probabilistic_analysis_functions.R
\name{generate_psa_params}
\alias{generate_psa_params}
\title{Generate PSA dataset of CEA parameters}
\usage{
generate_psa_params(n_sim = 1000, seed = 20190220)
}
\arguments{
\item{n_sim}{Number of PSA samples.}

\item{seed}{Seed for reproducibility of Monte Carlo sampling.}
}
\value{
A data frame with \code{n_sim} rows and 15 columns of parameters for PSA.
Each row is a parameter set sampled from distributions that characterize
their uncertainty
}
\description{
\code{generate_psa_params} generates PSA input dataset by sampling decision
model parameters from their distributions. The sample of the calibrated
parameters is a draw from their posterior distribution obtained with the
IMIS algorithm.
}
\examples{
generate_psa_params()
}
