% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_model_inputs_functions.R
\name{load_mort_data}
\alias{load_mort_data}
\title{Load mortality data}
\usage{
load_mort_data(file = NULL)
}
\arguments{
\item{file}{String with the location and name of the file with mortality
data. If \code{NULL}, \code{v_r_mort_by_age} will be used as default}
}
\value{
A vector with mortality by age.
}
\description{
\code{load_mort_data} is used to load age-specific mortality from .csv file
into vector.
}
