% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_guide.R
\name{open_guide}
\alias{open_guide}
\title{Open UserGuide of the package}
\usage{
open_guide(ext = "html")
}
\arguments{
\item{ext}{extension of the book to open: 'html', 'pdf'}
}
\description{
Open UserGuide of the package
}
