% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_calibration_functions.R
\name{sample.prior}
\alias{sample.prior}
\title{Sample from prior distributions of calibrated parameters}
\usage{
sample.prior(n_samp)
}
\arguments{
\item{n_samp}{Number of samples.}
}
\value{
A matrix with 3 rows and \code{n_samp} rows. Each row corresponds to a 
parameter set sampled from their prior distributions
}
\description{
\code{sample.prior} generates a sample of parameter sets from their prior 
distribution.
}
\examples{
v_param_names  <- c("p_S1S2", "hr_S1", "hr_S2")
n_param        <- length(v_param_names)
v_lb <- c(p_S1S2 = 0.01, hr_S1 = 1.0, hr_S2 = 5)  # lower bound
v_ub <- c(p_S1S2 = 0.50, hr_S1 = 4.5, hr_S2 = 15) # upper bound
sample.prior(2)
}
