% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_model_inputs_functions.R
\name{update_param_list}
\alias{update_param_list}
\title{Update parameters}
\usage{
update_param_list(l_params_all, params_updated)
}
\arguments{
\item{l_params_all}{List with all parameters of decision model}

\item{params_updated}{Parameters for which values need to be updated}
}
\value{
A list with all parameters updated.
}
\description{
\code{update_param_list} is used to update list of all parameters with new
values for specific parameters.
}
